charliecloud: add bind-mounts and container env
currently charliecloud can't bind mount directories that don't
exist within the container
This adds a workaround uses mkdir to create them before
bind-mounting.
ref: https://github.com/hpc/charliecloud/issues/96

Also, ENV layers are not honored by charliecloud when images are
pulled from a docker registry using 'ch-grow pull'
When the container contains software outside of the standard $PATH
of the host, it needs to be appended manually.
This workaround assumes that containers have the file
'/etc/environment' present, which can be used with
ch-run --set-env
ref:https://github.com/hpc/charliecloud/issues/719

TODO:
 * handle additional environment defined in nextflow env scope
 * extend mkdir workaround to additional mounts
 * add tests

This is a continuation of
nextflow-io#1712 by @tbugfinder

Signed-off-by: Patrick Hüther <patrick.huether@gmail.com>